Diverse rydding/refactor
